RESTful API/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
sqlalchemy-utils
alembic
